requirements.txt
langchain-core>=0.2.0
langchain-google-genai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langgraph>=0.2.0
chromadb>=0.4.0
pypdf>=4.0.0
python-dotenv>=1.0.0 